fix(policies): make options for policies simpler (#4722)
- We were exposing all the parameters that were introduced
for transparent back compat with old policies and resources.
As everything is meant to be standard with these new policies
it's not required to expose all this and we should have a small
API with little tweakable parameters
- make Validate a separate runtime checked method

+++ tools/policy-gen/generate-policy-import.sh
@@ -1,19 +1,13 @@
 #!/bin/bash
 
-IMPORTS=""
-for i in "$@"; do
-  if [[ -f pkg/plugins/policies/${i}/plugin.go ]]; then
-    IMPORTS="${IMPORTS}\t_ \"github.com/kumahq/kuma/pkg/plugins/policies/${i}\"\n"
-  fi
-done
-
-rm -f pkg/plugins/policies/imports.go
-if [[ ${IMPORTS} == "" ]]; then
+imports=$(for i in "$@"; do [[ -f pkg/plugins/policies/${i}/plugin.go ]] && echo "_ \"github.com/kumahq/kuma/pkg/plugins/policies/${i}\""; done)
+if [[ $imports == "" ]]; then
+  rm -f pkg/plugins/policies/imports.go
   exit 0
 fi
+
 echo "package policies
 
 import (
-${IMPORTS}
 )
 " > pkg/plugins/policies/imports.go
